python/requirements.txt
eth-erc20~=0.7.3
chainlib-eth~=0.4.18
